Add test for nested template literals
Initially I thought this package wouldn't work for nested template
literals (which ChromeDevTools used). I wanted to do TDD and add
the test first (since there was no test for it yet) and I discovered
this use case is already covered. That's because the `template.tag`
for MemberExpressions already includes the whole line, and thus
it matches on `lihtml.html`, which indeed includes `html`.

So let's add the test to make sure this use case remains covers,
but no actual code changes are required :tada: